# application/api.yaml
openapi: 3.0.1
info:
  description: Api Documentation
  version: '1.0'
  title: Api Documentation
  termsOfService: 'urn:tos'
  contact: {}
paths:
...
  /groups:
    get:
      summary: Gets the full list of groups
      operationId: groupsBrowse
      security:
      - authCookie: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Group'
    post:
      summary: Creates a new group
      operationId: groupsAdd
      security:
      - authCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Group'
      responses:
        '201':
          description: Group has been created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
        '400':
          description: Bad Request
        '401':
          description: User must be logged in to create new group
        '403':
          description: User is not allowed to create a new group
    put:
      summary: Edits an existing group
      operationId: groupsEdit
      security:
      - authCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Group'
      responses:
        '200':
          description: Group has been modified
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
        '400':
          description: Bad Request
        '401':
          description: User must be logged in to edit group
        '403':
          description: User is not allowed to edit this group
  /users:
    get:
      summary: Gets the full list of users
      operationId: usersBrowse
      security:
      - authCookie: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
    post:
      summary: Creates a new user
      operationId: usersAdd
      security:
      - authCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '201':
          description: User has been created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Bad Request
        '401':
          description: User must be logged in to create new user
        '403':
          description: User is not allowed to create a new user
  '/users/{userId}':
    get:
      summary: Gets the specified user
      operationId: usersRead
      security:
      - authCookie: []
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: User must be logged in to see this user
        '403':
          description: User is not allowed to see this user
        '404':
          description: User does not exist
  '/runner/{exerciseId}':
    post:
      deprecated: true
      summary: Runs provided code
      operationId: runCodeForExercise
      parameters:
      - name: exerciseId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RunnerArguments'
      responses:
        '200':
          description: OK
        '400':
          $ref: '#/components/responses/RFC7807Error'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
...
servers:
- url: '//localhost:8080/'
components:
  schemas:
    RunnerArguments:
      type: object
      properties:
        code:
          description: code
          type: string
        language:
          description: language
          type: string
      example:
        language: js
        code: 'function main() { console.log(''Hello World''); }'
...
    RFC7807Body:
      type: object
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        detail:
          type: string
        status:
          type: integer
          format: int64
        instance:
          type: string
          format: uri
      xml:
        name: problem
...
    User:
      type: object
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
      required:
      - id
      - firstName
      - lastName
      - email
      - password
...
    Semesters:
      description: Semester values.
      type: string
      enum:
      - FALL
      - WINTER
      - SUMMER
...
    Group:
      description: Group of students
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        groupNumber:
          type: integer
        class:
          type: string
        semester:
          $ref: '#/components/schemas/Semesters'
        year:
          type: number
        students:
          type: array
          items:
            $ref: '#/components/schemas/Student'
        archived:
          type: boolean
      required:
      - id
      - groupNumber
      - class
      - semester
      - year
      - students
      - archived
    Student:
      description: A student user
      type: object
      properties:
        accessCode:
          type: integer
        firstName:
          type: integer
        lastName:
          type: string
        email:
          type: string
        totalExercisesSuccessed:
          type: number
        totalExercisesDone:
          type: number
      required:
      - accessCode
      - firstName
      - lastName
      - type
  securitySchemes:
    authCookie:
      type: apiKey
      in: cookie
      name: JSESSIONID
  responses:
    RFC7807Error:
      description: Api problem response
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/RFC7807Body'
        application/problem+xml:
          schema:
            $ref: '#/components/schemas/RFC7807Body'