# application/generated/src/main/java/com/etsmtl/codecrusade/controllers/GroupsApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (3.3.2).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.etsmtl.codecrusade.controllers;

import com.etsmtl.codecrusade.model.Group;
import io.swagger.annotations.*;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RequestPart;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
@javax.annotation.Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2018-11-12T18:01:20.447505100-05:00[America/New_York]")

@Validated
@Api(value = "groups", description = "the groups API")
public interface GroupsApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    @ApiOperation(value = "Creates a new group", nickname = "groupsAdd", notes = "", response = Group.class, authorizations = {
        @Authorization(value = "authCookie")
    }, tags={  })
    @ApiResponses(value = { 
        @ApiResponse(code = 201, message = "Group has been created", response = Group.class),
        @ApiResponse(code = 400, message = "Bad Request"),
        @ApiResponse(code = 401, message = "User must be logged in to create new group"),
        @ApiResponse(code = 403, message = "User is not allowed to create a new group") })
    @RequestMapping(value = "/groups",
        produces = { "application/json" }, 
        consumes = { "application/json" },
        method = RequestMethod.POST)
    default ResponseEntity<Group> groupsAdd(@ApiParam(value = "" ,required=true )  @Valid @RequestBody Group group) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    ApiUtil.setExampleResponse(request, "application/json", "{  \"archived\" : true,  \"year\" : 1.46581298050294517310021547018550336360931396484375,  \"students\" : [ {    \"totalExercisesSuccessed\" : 2.3021358869347654518833223846741020679473876953125,    \"firstName\" : 5,    \"lastName\" : \"lastName\",    \"accessCode\" : 5,    \"totalExercisesDone\" : 7.061401241503109105224211816675961017608642578125,    \"email\" : \"email\"  }, {    \"totalExercisesSuccessed\" : 2.3021358869347654518833223846741020679473876953125,    \"firstName\" : 5,    \"lastName\" : \"lastName\",    \"accessCode\" : 5,    \"totalExercisesDone\" : 7.061401241503109105224211816675961017608642578125,    \"email\" : \"email\"  } ],  \"id\" : 0,  \"groupNumber\" : 6,  \"class\" : \"class\"}");
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    @ApiOperation(value = "Gets the full list of groups", nickname = "groupsBrowse", notes = "", response = Group.class, responseContainer = "List", authorizations = {
        @Authorization(value = "authCookie")
    }, tags={  })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "OK", response = Group.class, responseContainer = "List") })
    @RequestMapping(value = "/groups",
        produces = { "application/json" }, 
        method = RequestMethod.GET)
    default ResponseEntity<List<Group>> groupsBrowse() {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    ApiUtil.setExampleResponse(request, "application/json", "{  \"archived\" : true,  \"year\" : 1.46581298050294517310021547018550336360931396484375,  \"students\" : [ {    \"totalExercisesSuccessed\" : 2.3021358869347654518833223846741020679473876953125,    \"firstName\" : 5,    \"lastName\" : \"lastName\",    \"accessCode\" : 5,    \"totalExercisesDone\" : 7.061401241503109105224211816675961017608642578125,    \"email\" : \"email\"  }, {    \"totalExercisesSuccessed\" : 2.3021358869347654518833223846741020679473876953125,    \"firstName\" : 5,    \"lastName\" : \"lastName\",    \"accessCode\" : 5,    \"totalExercisesDone\" : 7.061401241503109105224211816675961017608642578125,    \"email\" : \"email\"  } ],  \"id\" : 0,  \"groupNumber\" : 6,  \"class\" : \"class\"}");
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    @ApiOperation(value = "Edits an existing group", nickname = "groupsEdit", notes = "", response = Group.class, authorizations = {
        @Authorization(value = "authCookie")
    }, tags={  })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "Group has been modified", response = Group.class),
        @ApiResponse(code = 400, message = "Bad Request"),
        @ApiResponse(code = 401, message = "User must be logged in to edit group"),
        @ApiResponse(code = 403, message = "User is not allowed to edit this group") })
    @RequestMapping(value = "/groups",
        produces = { "application/json" }, 
        consumes = { "application/json" },
        method = RequestMethod.PUT)
    default ResponseEntity<Group> groupsEdit(@ApiParam(value = "" ,required=true )  @Valid @RequestBody Group group) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    ApiUtil.setExampleResponse(request, "application/json", "{  \"archived\" : true,  \"year\" : 1.46581298050294517310021547018550336360931396484375,  \"students\" : [ {    \"totalExercisesSuccessed\" : 2.3021358869347654518833223846741020679473876953125,    \"firstName\" : 5,    \"lastName\" : \"lastName\",    \"accessCode\" : 5,    \"totalExercisesDone\" : 7.061401241503109105224211816675961017608642578125,    \"email\" : \"email\"  }, {    \"totalExercisesSuccessed\" : 2.3021358869347654518833223846741020679473876953125,    \"firstName\" : 5,    \"lastName\" : \"lastName\",    \"accessCode\" : 5,    \"totalExercisesDone\" : 7.061401241503109105224211816675961017608642578125,    \"email\" : \"email\"  } ],  \"id\" : 0,  \"groupNumber\" : 6,  \"class\" : \"class\"}");
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}